refactoring mains
