+++ src/cpp/CMakeLists.txt
@@ -18,5 +18,6 @@ add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/lpnamer")
 add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/benders")
 add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/multisolver_interface")
 add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/sensitivity")
+add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/full_run")
 
 add_subdirectory ("${CMAKE_CURRENT_SOURCE_DIR}/exe")
