+++ src/cpp/lpnamer/main/ProblemGenerationExeOptions.cpp
@@ -1,0 +1,14 @@
+#include "ProblemGenerationExeOptions.h"
+namespace po = boost::program_options;
+
+ProblemGenerationExeOptions::ProblemGenerationExeOptions() {
+  AddOptions()("help,h", "produce help message")(
+      "output,o", po::value<std::filesystem::path>(&root_)->required(),
+      "antares-xpansion study output")(
+      "formulation,f",
+      po::value<std::string>(&master_formulation_)->default_value("relaxed"),
+      "master formulation (relaxed or integer)")(
+      "exclusion-files,e",
+      po::value<std::string>(&additional_constraintFilename_l_),
+      "path to exclusion files");
+}
